+++ modules/evaluation-server/tests/Domain.UnitTests/Domain.UnitTests.csproj
@@ -14,7 +14,6 @@
 
     <ItemGroup>
         <ProjectReference Include="..\TestBase\TestBase.csproj" />
-        <ProjectReference Include="..\..\src\Domain\Domain.csproj" />
     </ItemGroup>
 
 </Project>
